using tweetService as the main file for tweet api calls
