+++ src/services/tweetService.js
@@ -1,4 +1,5 @@
 import axios from 'axios';
+import { formatDate, formatDateForComments } from '../utils/dateUtils';
 
 export const postTweet = async (tweetText) => {
   try {
@@ -6,6 +7,49 @@ export const postTweet = async (tweetText) => {
       'http://localhost:3003/api/v1/tweets',
       {
         texto: tweetText,
+      },
+      {
+        headers: {
+          Authorization: `Bearer ${localStorage.getItem('token')}`,
+        },
+      },
+    );
+    return response.data;
+  } catch (error) {
+    console.error(error);
+  }
+};
+
+export const getTweetDetails = async (id) => {
+  try {
+    const response = await axios.get(
+      `http://localhost:3003/api/v1/tweets/${id}`,
+      {
+        headers: {
+          Authorization: `Bearer ${localStorage.getItem('token')}`,
+        },
+      },
+    );
+
+    const tweetDetails = response.data;
+    tweetDetails.conteudo.data = formatDate(tweetDetails.conteudo.data);
+
+    tweetDetails.conteudo.comentariosLista.forEach((commentTweet) => {
+      commentTweet.data = formatDateForComments(commentTweet.data);
+    });
+
+    return tweetDetails;
+  } catch (error) {
+    console.error('Erro ao buscar Tweets:', error);
+  }
+};
+
+export const postComment = async (commentText, tweetId) => {
+  try {
+    const response = await axios.post(
+      `http://localhost:3003/api/v1/tweets/${tweetId}/comentarios`,
+      {
+        texto: commentText,
       },
       {
         headers: {
